+++ src/api/Action.ts
@@ -57,9 +57,8 @@ export interface Action {
      *
      * @param name The name of the error.
      * @param code The error code.
-     * @param message The message (reason) of the error.
      */
-    reportError(name: string, code: number, message: string): void;
+    reportError(name: string, code: number): void;
 
     /**
      * Leave the action.