+++ src/core/impl/null/NullSession.ts
@@ -32,7 +32,7 @@ export class NullSession implements Session {
         // stub
     }
 
-    public reportError(name: string, code: number, message: string): void {
+    public reportError(name: string, code: number): void {
         // stub
     }